refactor: decouple ingest function from ingest script
